Blueprints added
