+++ Dockerfile
@@ -5,7 +5,6 @@ WORKDIR /app
 
 ENV PYTHONDONTWRITEBYTECODE 1
 ENV PYTHONUNBUFFERED 1
-ENV FLASK_APP=/app/fxprofile
 ENV HOST=0.0.0.0
 
 RUN pip install -r requirements.txt
